+++ sdk/formrecognizer/azure-ai-formrecognizer/src/samples/java/com/azure/ai/formrecognizer/AnalyzeIdentityDocumentsFromUrl.java
@@ -36,7 +36,7 @@ public class AnalyzeIdentityDocumentsFromUrl {
 
         String licenseDocumentUrl =
             "https://raw.githubusercontent.com/Azure/azure-sdk-for-java/main/sdk/formrecognizer/"
-                + "azure-ai-formrecognizer/src/samples/resources/sample-forms/IdentityDocuments/license.jpg";
+                + "azure-ai-formrecognizer/src/samples/resources/sample-forms/IdentityDocuments/license.png";
         SyncPoller<DocumentOperationResult, AnalyzeResult> analyzeIdentityDocumentPoller
             = client.beginAnalyzeDocumentFromUrl("prebuilt-idDocument", licenseDocumentUrl);